fix: ID sometimes not a string, use a new function to solve it.
